django/requirements.txt
asgiref==3.4.1
bioformats==0.1.15
boto3==1.18.57
botocore==1.21.57
cycler==0.10.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
imageio==2.9.0
isort==5.9.3
jmespath==0.10.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
matplotlib==3.4.3
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
platformdirs==2.4.0
pyfaidx==0.6.2
pylint==2.11.1
pyparsing==2.4.7

python-dateutil==2.8.2

pytz==2021.3
pyvcf==0.4.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.1

celery==4.2.0